+++ test/org/usfirst/team1699/test/autonomous/commands/Shoot.java
@@ -17,7 +17,7 @@ public class Shoot extends org.usfirst.frc.team1699.utils.command.Command{
 	}
 	
 	@Override
-	public void runAuto(int distance, double speed, boolean useSensor) {
+	public void runAuto(double distance, double speed, boolean useSensor) {
 		System.out.println(getName() + " " + distance + " " + speed + " " + useSensor);
 	}
